refactor: break up `demo-server` into modules
